Modificando funcionalidd endpoind requerimiento 2, exception

+++ src/main/java/com/bootcamp/finalProject/dtos/OrderRequestDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 
 public class OrderRequestDTO {
 
-    private Integer dealerNumber;
+    private Long dealerNumber;
     private String deliveryStatus;
     private Integer order;
 }
